Addition of spring supports #5 , data validation enhancement #6 , trapezoidal load #4
Major update to indeterminatebeam.py with better documentation, that is consistently in accordance with numpy style docstrings. Have also added a simple wrapper for trapezoidal loading which is defined as a function that creates a DistributedLoadV class. Have also implemented spring supports, which involved chaning the support class to have a stiffness parameter and changing the analyse function. The function was also generally tidied and improved to be shorter. I also added two functions for data validation that help verify the user has correcly input values. These functions are currently stored in a different file called data_validation. The documentation was slightly updated to include the theory behind the spring supports analysis method.