DistributedLoad Update
Add DistributedLoad that takes angle to transpose into horizontal and vertical distributed loads. The positions of the distributed load are fixed and dont transpose, rather it is the direction of the arrows.
Plotting was established for DistributedLoad and DistributedLoadH load types and colors were chosen to be more clear, which involved changing the current color for DistributedLoadV.