Loading, analysis, plotting update for Singularity
Modified Analysis to work with singularity function, updated plotting (except for reaction plot) to work with singularity function solution. Cant use lambdify anymore so plotting is now slower than before but the analysis is much faster and the code is a lot more readable so it is a good compromise. Moved loading classes into a seperate file in order to allow plot_drawing_aid to import loading classes without causing circular dependency issues.