PEP8 update and spring bug fix
Checked and modified files to be to pep8 standard using autopep8 and pep8online checker. In some instances left long lines. Also fixed a bug for springs in the analysis section which involved changing unknown['force'] to unknown['variable'] which simply meant that a singularity function wasnt added. (The singularity function was never needed since we were always evaluating at the point of the 'singularity'.)