Timoshenko beam theory
Changed the deflection definition from Euler to Timoshenko beam theory in order to account for shear deformation. Since the program integrates from the internal forces to find the deformations, the only difference is modifying v = integral(phi) to v = integral(phi + V/GA).

For this new addition, the shear modulus (G) was added as a beam parameter. This has been updated in units.py as well